+++ Runtime/SDK/APIHandler/PixoPlatformAPIHandler.cs
@@ -196,7 +196,7 @@ namespace PixoVR.Apex
             success?.Invoke(response, responseContent);
         }
 
-        public override async void GetUserMetricsForOrg(string authToken, int orgID, int page, FilterParams filterParams, Action<UserMetricsResponse, object> success, Action<HttpResponseMessage, FailureResponse> failure)
+        public override async void GetUserMetricsForOrg(string authToken, int orgID, int page, FilterParams filterParams, Action<HttpResponseMessage, UserMetricsResponse> success, Action<HttpResponseMessage, FailureResponse> failure)
         {
             var paramsInput = new
             {
@@ -226,7 +226,6 @@ namespace PixoVR.Apex
                     var failureResponse = GetGQLFailureResponse(jsonResponse, "userMetrics");
                     if (failureResponse != null)
                     {
-                        OnAPIResponse?.Invoke(ResponseType.RT_GET_USER_METRICS_FOR_ORG, response, failureResponse);
                         failure?.Invoke(response, failureResponse);
                         return;
                     }
@@ -235,7 +234,7 @@ namespace PixoVR.Apex
                     var userMetricsResponse = JsonConvert.DeserializeObject<UserMetricsResponse>(userMetricsJSON.ToString());
                     userMetricsResponse.result.ForEach(u => u.RefreshDisplayFields());
 
-                    success?.Invoke(userMetricsResponse, response);
+                    success?.Invoke(response, userMetricsResponse);
                 }
                 catch (Exception ex)
                 {